+++ src/syntax/_color-palette.js
@@ -1,5 +1,3 @@
-"use strict";
-
 const syntaxColors = {
 	Normal: {
 		pink: "#FF9EE7",
@@ -10,6 +8,7 @@ const syntaxColors = {
 		green: "#2FDE73",
 		yellow: "#E6E667",
 		white: "#D3D1EB",
+		grey: "#7F848C",
 		red: "#E6457B",
 		brown: "#99611F",
 	},
@@ -22,6 +21,7 @@ const syntaxColors = {
 		green: "#29BD62",
 		yellow: "#CBCB5B",
 		white: "#BBBAD2",
+		grey: "#71767D",
 		red: "#CC3D69",
 		brown: "#82511A",
 	},
@@ -74,6 +74,7 @@ const schemes = (options, colors) => {
 		number: colors.pink,
 		variable: colors.white,
 		punctuation: colors.cyan,
+		grey: colors.grey,
 		invalid: colors.red,
 		deprecated: colors.brown,
 	};
@@ -86,6 +87,7 @@ const schemes = (options, colors) => {
 		number: colors.yellow,
 		variable: colors.white,
 		punctuation: colors.cyan,
+		grey: colors.grey,
 		invalid: colors.red,
 		deprecated: colors.brown,
 	};
@@ -98,6 +100,7 @@ const schemes = (options, colors) => {
 		number: colors.pink,
 		variable: colors.white,
 		punctuation: colors.cyan,
+		grey: colors.grey,
 		invalid: colors.red,
 		deprecated: colors.brown,
 	};
@@ -110,6 +113,7 @@ const schemes = (options, colors) => {
 		number: colors.yellow,
 		variable: colors.white,
 		punctuation: colors.cyan,
+		grey: colors.grey,
 		invalid: colors.red,
 		deprecated: colors.brown,
 	};