+++ src/syntax/syntax-textmate.js
@@ -1,5 +1,3 @@
-"use strict";
-
 const textmateTokens = (colorSchemes, schemeName) => {
 	const scheme = colorSchemes[schemeName];
 	return [
@@ -43,7 +41,11 @@ const textmateTokens = (colorSchemes, schemeName) => {
 			},
 		},
 		{
-			scope: ["variable.language.this", "variable.language.self"],
+			scope: [
+				"variable.language.this",
+				"variable.language.self",
+				"keyword.other.important",
+			],
 			settings: {
 				foreground: scheme.class,
 			},
@@ -55,11 +57,7 @@ const textmateTokens = (colorSchemes, schemeName) => {
 			},
 		},
 		{
-			scope: [
-				"keyword.control",
-				"punctuation.definition.keyword",
-				"keyword.other.important",
-			],
+			scope: ["keyword.control", "punctuation.definition.keyword"],
 			settings: {
 				foreground: scheme.storageType,
 			},
@@ -501,7 +499,6 @@ const textmateTokens = (colorSchemes, schemeName) => {
 				"punctuation.definition.arguments.end",
 				"punctuation.definition.entity.begin",
 				"punctuation.definition.entity.end",
-				"punctuation.definition.tag",
 				"punctuation.definition.type.begin",
 				"punctuation.definition.type.end",
 				"punctuation.section.scope.begin",
@@ -512,6 +509,12 @@ const textmateTokens = (colorSchemes, schemeName) => {
 			],
 			settings: {
 				foreground: scheme.variable,
+			},
+		},
+		{
+			scope: ["punctuation.definition.tag"],
+			settings: {
+				foreground: scheme.grey,
 			},
 		},
 		{
@@ -551,22 +554,31 @@ const textmateTokens = (colorSchemes, schemeName) => {
 			},
 		},
 		{
-			scope: ["entity.other.attribute-name"],
+			scope: ["entity.name.tag.wildcard"],
+			settings: {
+				foreground: scheme.punctuation,
+			},
+		},
+		{
+			scope: [
+				"entity.other.attribute-name",
+				"entity.other.attribute-name.id",
+			],
 			settings: {
 				foreground: scheme.number,
 			},
 		},
 		{
-			scope: ["entity.other.attribute-name.id"],
+			scope: [
+				"entity.other.attribute-name.pseudo-class",
+				"entity.other.attribute-name.pseudo-element",
+			],
 			settings: {
 				foreground: scheme.function,
 			},
 		},
 		{
 			scope: [
-				"entity.other.attribute-name.pseudo-element",
-				"entity.other.attribute-name.pseudo-class",
-				"entity.other.attribute-name.class",
 				"entity.other.attribute-name.class.mixin",
 				"entity.other.attribute-name.parent-selector",
 				"entity.other.attribute-name.attribute",
@@ -576,6 +588,15 @@ const textmateTokens = (colorSchemes, schemeName) => {
 			],
 			settings: {
 				foreground: scheme.number,
+			},
+		},
+		{
+			scope: [
+				"entity.other.attribute-name.class",
+				"meta.attribute.class.html entity.other.attribute-name.html",
+			],
+			settings: {
+				foreground: scheme.storageType,
 			},
 		},
 		{
@@ -594,7 +615,7 @@ const textmateTokens = (colorSchemes, schemeName) => {
 		{
 			scope: ["support.constant.property-value"],
 			settings: {
-				foreground: scheme.variable,
+				foreground: scheme.property,
 			},
 		},
 		{
@@ -647,6 +668,15 @@ const textmateTokens = (colorSchemes, schemeName) => {
 			],
 			settings: {
 				foreground: scheme.storageType,
+			},
+		},
+		{
+			scope: [
+				"punctuation.definition.annotation.java",
+				"punctuation.definition.decorator.python",
+			],
+			settings: {
+				foreground: scheme.class,
 			},
 		},
 		{